+++ README.md
@@ -38,9 +38,10 @@ pnpm add -D solid-marked
 
 ## Usage
 
-### Plugins
+### Integrations
 
-- [Vite](https://github.com/LXSMNSYC/solid-marked/tree/main/packages/vite-plugin-solid-marked)
+- [Unplugin](https://github.com/LXSMNSYC/solid-marked/tree/main/packages/unplugin)
+- [Vite](https://github.com/LXSMNSYC/solid-marked/tree/main/packages/vite)
 
 ### Using the compiler
 
@@ -48,37 +49,49 @@ pnpm add -D solid-marked
 import { compile } from 'solid-marked';
 
 const { map, code } = await compile(
-  'mdx-provider',  // Where to import the builtin components
   'my-file.md',    // Name of the file
   '# Hello World', // Markdown code
+  {
+    // Where to import the builtin components
+    // Default: solid-marked
+    mdxImportSource: 'mdx-provider', 
+    // If solid-marked should use the Dynamic component
+    // or not.
+    // Possible values:
+    // - true: completely disables Dynamic component
+    // - false: enables Dynamic component
+    // - 'only-mdx': Dynamic component is only enabled for MDX
+    noDynamicComponents: 'only-mdx',
+  }
 );
 
 console.log(code);
-
-// Output:
-// import { useMDX } from 'mdx-provider';
-
-// export default function Component(props) {
-//   const __ctx = useMDX();
-//   return (
-//     <Dynamic component={__ctx.builtins.Root}>
-//       <Dynamic component={__ctx.builtins.Heading} depth={1}>Hello World</Dynamic>
-//     </Dynamic>
-//   );
-// }
 ```
 
-### `useMDX`
+### `useMDX` and `MDXProvider`
 
-Components generated by `solid-marked` uses the fundamental components from an MDX provider, this is through the use of `useMDX` which is imported from the module.
+Components generated by `solid-marked` uses the fundamental components from an MDX provider, this is through the use of `useMDX` which is imported from the module. By default, `useMDX` comes from `solid-marked` and needs to be used in conjunction with `<MDXProvider>`
 
-### Typescript
+```js
+import { MDXProvider } from 'solid-marked';
 
-```ts
-/// <reference types="solid-marked/env">
+// ...
+<MDXProvider
+  builtins={{
+    Heading(props) {
+      return (
+        <Dynamic component={`h${props.depth}`}>
+          {props.children}
+        </Dynamic>
+      );
+    },
+  }}
+>
+  <App />
+</MDXProvider>
 ```
 
-Example module
+Example custom module
 
 ```js
 export function useMDX() {
@@ -92,6 +105,12 @@ export function useMDX() {
     },
   };
 }
+```
+
+### Typescript
+
+```ts
+/// <reference types="solid-marked/env">
 ```
 
 ## Sponsors
