+++ examples/vite-demo/.eslintrc.cjs
