+++ examples/vite-demo/package.json
@@ -1,22 +1,23 @@
 {
   "name": "vite-demo",
+  "type": "module",
   "version": "0.4.1",
   "devDependencies": {
-    "@tailwindcss/typography": "^0.5.4",
-    "autoprefixer": "^10.4.7",
-    "eslint": "^8.20.0",
-    "eslint-config-lxsmnsyc": "^0.4.7",
-    "postcss": "^8.4.14",
+    "@tailwindcss/typography": "^0.5.9",
+    "autoprefixer": "^10.4.14",
+    "eslint": "^8.38.0",
+    "eslint-config-lxsmnsyc": "^0.5.1",
+    "postcss": "^8.4.21",
     "solid-marked": "0.4.0",
-    "typescript": "^4.7.4",
-    "vite": "^2.9.12",
-    "vite-plugin-solid": "2.2.6",
-    "vite-plugin-solid-marked": "0.4.1"
+    "typescript": "^4.9.5",
+    "vite": "^4.2.1",
+    "vite-plugin-solid": "2.7.0",
+    "unplugin-solid-marked": "0.4.1"
   },
   "private": true,
   "dependencies": {
-    "solid-js": "^1.4.7",
-    "tailwindcss": "^3.1.6"
+    "solid-js": "^1.7.3",
+    "tailwindcss": "^3.3.1"
   },
   "scripts": {
     "dev": "vite",
