+++ examples/vite-demo/tailwind.config.cjs
