+++ examples/vite-demo/vite.config.ts
@@ -1,12 +1,10 @@
 import { defineConfig } from 'vite';
 import solidPlugin from 'vite-plugin-solid';
-import solidMarkedPlugin from 'vite-plugin-solid-marked';
+import solidMarkedPlugin from 'unplugin-solid-marked';
 
 export default defineConfig({
   plugins: [
-    solidMarkedPlugin({
-      source: './markdown'
-    }),
+    solidMarkedPlugin.vite({}),
     solidPlugin(),
   ],
 });
