+++ package.json
@@ -6,9 +6,9 @@
     "examples/*"
   ],
   "devDependencies": {
-    "eslint": "^8.19.0",
-    "eslint-config-lxsmnsyc": "^0.4.7",
-    "lerna": "^5.1.8",
-    "typescript": "^4.7.4"
+    "eslint": "^8.38.0",
+    "eslint-config-lxsmnsyc": "^0.5.1",
+    "lerna": "^6.6.1",
+    "typescript": "^4.9.5"
   }
 }
