+++ packages/solid-marked/compiler/interfaces.ts
