+++ packages/solid-marked/package.json
@@ -1,18 +1,28 @@
 {
   "name": "solid-marked",
+  "type": "module",
   "version": "0.4.0",
-  "types": "dist/types/index.d.ts",
-  "main": "dist/cjs/production/index.js",
-  "module": "dist/esm/production/index.js",
+  "types": "./dist/types/src/index.d.ts",
+  "main": "./dist/cjs/production/index.cjs",
+  "module": "./dist/esm/production/index.mjs",
   "exports": {
     ".": {
       "development": {
-        "require": "./dist/cjs/development/index.js",
-        "import": "./dist/esm/development/index.js"
+        "require": "./dist/cjs/development/index.cjs",
+        "import": "./dist/esm/development/index.mjs"
       },
-      "require": "./dist/cjs/production/index.js",
-      "import": "./dist/esm/production/index.js",
-      "types": "./dist/types/index.d.ts"
+      "require": "./dist/cjs/production/index.cjs",
+      "import": "./dist/esm/production/index.mjs",
+      "types": "./dist/types/src/index.d.ts"
+    },
+    "./compiler": {
+      "development": {
+        "require": "./dist/cjs/development/compiler.cjs",
+        "import": "./dist/esm/development/compiler.mjs"
+      },
+      "require": "./dist/cjs/production/compiler.cjs",
+      "import": "./dist/esm/production/compiler.mjs",
+      "types": "./dist/types/compiler/index.d.ts"
     }
   },
   "files": [
@@ -29,17 +39,17 @@
   ],
   "devDependencies": {
     "@types/mdast": "^3.0.10",
-    "@types/node": "^17.0.35",
-    "eslint": "^8.19.0",
-    "eslint-config-lxsmnsyc": "^0.4.7",
-    "pridepack": "2.2.1",
-    "solid-js": "^1.4.7",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.4",
-    "vitest": "^0.18.1"
+    "@types/node": "^18.15.11",
+    "eslint": "^8.38.0",
+    "eslint-config-lxsmnsyc": "^0.5.1",
+    "pridepack": "2.4.4",
+    "solid-js": "^1.7.3",
+    "tslib": "^2.5.0",
+    "typescript": "^4.9.5",
+    "vitest": "^0.30.1"
   },
   "peerDependencies": {
-    "solid-js": "^1.2"
+    "solid-js": "^1.7"
   },
   "dependencies": {
     "mdast-util-from-markdown": "^1.2.0",
@@ -76,8 +86,8 @@
   },
   "typesVersions": {
     "*": {
-      "*": [
-        "./dist/types/index.d.ts"
+      "compiler": [
+        "./dist/types/compiler/index.d.ts"
       ]
     }
   }
