+++ packages/solid-marked/pridepack.json
@@ -1,3 +1,7 @@
 {
-  "target": "es2017"
+  "target": "es2017",
+  "entrypoints": {
+    ".": "src/index.ts",
+    "./compiler": "compiler/index.ts"
+  }
 }
