+++ packages/unplugin/package.json
@@ -1,0 +1,82 @@
+{
+  "name": "unplugin-solid-marked",
+  "version": "0.4.1",
+  "type": "module",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "engines": {
+    "node": ">=10"
+  },
+  "license": "MIT",
+  "keywords": [
+    "pridepack"
+  ],
+  "devDependencies": {
+    "@types/node": "^18.15.11",
+    "eslint": "^8.38.0",
+    "eslint-config-lxsmnsyc": "^0.5.1",
+    "pridepack": "2.4.4",
+    "solid-js": "^1.7.3",
+    "solid-marked": "0.4.0",
+    "tslib": "^2.5.0",
+    "typescript": "^4.9.5",
+    "vite": "^4.2.1"
+  },
+  "peerDependencies": {
+    "solid-js": "^1.7",
+    "solid-marked": "0.4.0",
+    "vite": "^4"
+  },
+  "peerDependenciesMeta": {
+    "vite": {
+      "optional": true
+    }
+  },
+  "scripts": {
+    "prepublishOnly": "pridepack clean && pridepack build",
+    "build": "pridepack build",
+    "type-check": "pridepack check",
+    "lint": "pridepack lint",
+    "clean": "pridepack clean",
+    "watch": "pridepack watch",
+    "start": "pridepack start",
+    "dev": "pridepack dev",
+    "test": "vitest"
+  },
+  "private": false,
+  "description": "Vite plugin for solid-marked",
+  "repository": {
+    "url": "https://github.com/lxsmnsyc/solid-marked.git",
+    "type": "git"
+  },
+  "homepage": "https://github.com/lxsmnsyc/solid-marked/tree/main/packages/unplugin",
+  "bugs": {
+    "url": "https://github.com/lxsmnsyc/solid-marked/issues"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "author": "Alexis Munsayac",
+  "exports": {
+    ".": {
+      "development": {
+        "require": "./dist/cjs/development/index.cjs",
+        "import": "./dist/esm/development/index.mjs"
+      },
+      "require": "./dist/cjs/production/index.cjs",
+      "import": "./dist/esm/production/index.mjs",
+      "types": "./dist/types/index.d.ts"
+    }
+  },
+  "typesVersions": {
+    "*": {}
+  },
+  "dependencies": {
+    "unplugin": "^1.3.1"
+  },
+  "types": "./dist/types/index.d.ts",
+  "main": "./dist/cjs/production/index.cjs",
+  "module": "./dist/esm/production/index.mjs"
+}
