+++ packages/unplugin/pridepack.json
@@ -1,0 +1,3 @@
+{
+  "target": "es2018"
+}
