+++ packages/vite/package.json
@@ -1,6 +1,7 @@
 {
   "name": "vite-plugin-solid-marked",
   "version": "0.4.1",
+  "type": "module",
   "files": [
     "dist",
     "src"
@@ -13,23 +14,20 @@
     "pridepack"
   ],
   "devDependencies": {
-    "@types/node": "^18.0.0",
-    "eslint": "^8.18.0",
-    "eslint-config-lxsmnsyc": "^0.4.1",
-    "pridepack": "2.0.1",
-    "solid-js": "^1.4.7",
+    "@types/node": "^18.15.11",
+    "eslint": "^8.38.0",
+    "eslint-config-lxsmnsyc": "^0.5.1",
+    "pridepack": "2.4.4",
+    "solid-js": "^1.7.3",
     "solid-marked": "0.4.0",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.4",
-    "vite": "^3.0.3",
-    "vite-plugin-solid": "^2.3.0",
-    "vitest": "^0.16.0"
+    "tslib": "^2.5.0",
+    "typescript": "^4.9.5",
+    "vite": "^4.2.1"
   },
   "peerDependencies": {
     "solid-js": "^1.3",
-    "solid-marked": "^0.2",
-    "vite": "^2.0 || ^3.0",
-    "vite-plugin-solid": "^2.2"
+    "solid-marked": "0.4.0",
+    "vite": "^4.0"
   },
   "scripts": {
     "prepublishOnly": "pridepack clean && pridepack build",
@@ -48,7 +46,7 @@
     "url": "https://github.com/lxsmnsyc/solid-marked.git",
     "type": "git"
   },
-  "homepage": "https://github.com/lxsmnsyc/solid-marked/tree/main/packages/vite-plugin-solid-marked",
+  "homepage": "https://github.com/lxsmnsyc/solid-marked/tree/main/packages/unplugin",
   "bugs": {
     "url": "https://github.com/lxsmnsyc/solid-marked/issues"
   },
