+++ packages/vite/pridepack.json
