+++ packages/vite/tsconfig.eslint.json
@@ -1,6 +1,6 @@
 {
   "exclude": ["node_modules"],
-  "include": ["src", "types", "env.d.ts", "compiler"],
+  "include": ["src", "types", "test"],
   "compilerOptions": {
     "module": "ESNext",
     "lib": ["ESNext", "DOM"],
@@ -16,6 +16,6 @@
     "moduleResolution": "node",
     "jsx": "react",
     "esModuleInterop": true,
-    "target": "es2018"
+    "target": "ES2017"
   }
 }
